requirements.txt
aioredis==1.3.0
asgi-redis==1.0.0
asgiref==3.2.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==3.1.0
bs4==0.0.1
build==1.0.2
certifi==2019.9.11
cffi==1.13.2
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
Click==7.0
codecov==2.0.15
colorama==0.4.1
constantly==15.1.0
coverage==4.3.3
cryptography==2.8
cycler==0.10.0
Cython==0.29.14
daphne==0.15.0
decorator==4.4.1
defusedxml==0.6.0
descartes==1.1.0
django-extensions==2.2.5
django-model-utils==2.6.1
django-test-plus==1.0.17
docker==4.1.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.3.0
incremental==17.5.0
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.1
JPype1==0.7.0
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
kiwisolver==1.1.0
konlpy==0.5.2
lazy-object-proxy==1.4.3
lxml==4.4.2
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mecab-python===0.996-ko-0.9.2-msvc
missingno==0.4.2
mistune==0.8.4
mizani==0.6.0
mock==3.0.5
more-itertools==7.2.0
msgpack==0.6.2
msgpack-python==0.5.6
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.4+mkl
oauthlib==3.1.0
packaging==19.2
palettable==3.3.0
pandas==0.25.2
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
plotnine==0.6.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.4
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
PyHamcrest==1.9.0
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.2
pypiwin32==223
pyrsistent==0.15.4
PySocks==1.7.1
python-dateutil==2.8.0
python-decouple==3.3
python-dotenv==0.10.3
pytz==2019.3
pywin32==227
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.5
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
scikit-learn==0.22
scipy==1.3.1
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.10.0
sklearn==0.0
soupsieve==1.9.5
sqlparse==0.3.0
statsmodels==0.10.1
terminado==0.8.2
testpath==0.4.2
toml==0.10.0
tornado==6.0.3
tox==3.14.2
tqdm==4.40.2
traitlets==4.3.3
tweepy==3.8.0
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.7
virtualenv==16.7.9
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
zope.interface==4.7.1
